requirements.txt
acp-sdk==1.0.3
aiofile==3.9.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.95.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
async-lru==2.3.0
attrs==26.1.0
Authlib==1.6.11
azure-core==1.39.0
azure-storage-blob==12.28.0
beartype==0.22.9
cachetools==7.0.5
caio==0.9.25
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
cyclopts==4.10.2
distro==1.9.0
ddgs==9.14.1
dnspython==2.8.0
docstring_parser==0.18.0
docutils==0.22.4
email-validator==2.3.0
environs==9.5.0
exceptiongroup==1.3.1
fastapi==0.135.3
fastapi-cli==0.0.24
fastapi-cloud-cli==0.17.0
fastar==0.11.0
fastmcp==3.2.4
filelock==3.29.0
fsspec==2026.3.0
googleapis-common-protos==1.74.0
griffelib==2.0.2
grpcio==1.60.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
isodate==0.7.2
janus==2.0.0
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
keyring==25.7.0
langchain==1.2.15
langchain-anthropic==1.4.0
langchain-core==1.2.30
langchain-ollama==1.1.0
langgraph==1.1.6
langgraph-checkpoint==4.0.2
langgraph-checkpoint-postgres==3.0.5
langgraph-checkpoint-sqlite==3.0.3
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.32
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mcp==1.27.0
mdurl==0.1.2
minio==7.2.20
more-itertools==11.0.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
obstore==0.9.3
ollama==0.6.1
openapi-pydantic==0.5.1
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-http==1.41.0
opentelemetry-instrumentation==0.62b0
opentelemetry-instrumentation-asgi==0.62b0
opentelemetry-instrumentation-fastapi==0.62b0
opentelemetry-instrumentation-httpx==0.62b0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
opentelemetry-util-http==0.62b0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.1
pandas==3.0.2
pathable==0.5.0
pillow==12.2.0
platformdirs==4.9.6
protobuf==6.33.6
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg-pool==3.3.0
py-key-value-aio==0.4.4
pyarrow==24.0.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.13.1
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.46.1
Pygments==2.20.0
PyJWT==2.12.1
pymilvus==2.3.7
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
rich-rst==1.3.2
rich-toolkit==0.19.7
rignore==0.7.6
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==3.0.1
sentry-sdk==2.58.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlite-vec==0.1.9
sse-starlette==3.3.4
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==4.57.6
trafilatura==1.12.2
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.12.0
uncalled-for==0.3.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wheel==0.46.3
wrapt==2.1.2
xxhash==3.6.0
zipp==3.23.1
zstandard==0.25.0
